backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
gunicorn==21.2.0
motor==3.6.0
pydantic==2.7.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
boto3==1.34.81
pdfminer.six==20231228
python-docx==1.1.0
spacy==3.7.4
sentence-transformers==2.5.1
transformers==4.39.3
torch==2.2.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
python-multipart==0.0.9
tenacity==8.2.3
orjson==3.10.0
faker==19.13.0
slowapi==0.1.9
python-json-logger==2.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
httpx==0.27.0
pydantic[email]==2.7.0
email-validator

